sea orm 分支初始化

+++ src/vo/role_vo.rs
@@ -14,9 +14,9 @@ pub struct RoleListReq {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct RoleListData {
-    pub id: i32,
+    pub id: i64,
     pub sort: i32,
-    pub status_id: i32,
+    pub status_id: i8,
     pub role_name: String,
     pub remark: String,
     pub create_time: String,
@@ -28,16 +28,16 @@ pub struct RoleListData {
 pub struct RoleSaveReq {
     pub role_name: String,
     pub sort: i32,
-    pub status_id: i32,
+    pub status_id: i8,
     pub remark: Option<String>,
 }
 
 #[derive(Debug, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct RoleUpdateReq {
-    pub id: i32,
+    pub id: i64,
     pub sort: i32,
-    pub status_id: i32,
+    pub status_id: i8,
     pub role_name: String,
     pub remark: Option<String>,
 }
@@ -46,14 +46,14 @@ pub struct RoleUpdateReq {
 #[derive(Debug, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct RoleDeleteReq {
-    pub ids: Vec<i32>,
+    pub ids: Vec<i64>,
 }
 
 
 #[derive(Debug, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct QueryRoleMenuReq {
-    pub role_id: i32,
+    pub role_id: i64,
 }
 
 #[derive(Debug, Serialize)]
@@ -66,14 +66,14 @@ pub struct QueryRoleMenuResp {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct QueryRoleMenuData {
-    pub role_menus: Vec<i32>,
+    pub role_menus: Vec<i64>,
     pub menu_list: Vec<MenuDataList>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MenuDataList {
-    pub id: i32,
-    pub parent_id: i32,
+    pub id: i64,
+    pub parent_id: i64,
     pub title: String,
     pub key: String,
     pub label: String,
@@ -84,8 +84,8 @@ pub struct MenuDataList {
 #[derive(Debug, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct UpdateRoleMenuReq {
-    pub menu_ids: Vec<i32>,
-    pub role_id: i32,
+    pub menu_ids: Vec<i64>,
+    pub role_id: i64,
 }
